+++ cmd/evld/improve_genesis.go
@@ -1,0 +1,199 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"math/big"
+	"os"
+
+	sdkmath "cosmossdk.io/math"
+
+	cmttypes "github.com/cometbft/cometbft/types"
+	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
+
+	evmtypes "github.com/EscanBE/everlast/v12/x/evm/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	crisistypes "github.com/cosmos/cosmos-sdk/x/crisis/types"
+	govtypesv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
+	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+
+	"github.com/EscanBE/everlast/v12/constants"
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/codec"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	"github.com/spf13/cobra"
+)
+
+func NewImproveGenesisCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "improve-genesis",
+		Short: "Improve genesis by update the genesis.json file with necessary changes",
+		Args:  cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+			if homeDir := clientCtx.HomeDir; homeDir == "" {
+				return fmt.Errorf("home dir not set")
+			}
+
+			// Load the genesis file
+			genesisFile := fmt.Sprintf("%s/config/genesis.json", clientCtx.HomeDir)
+			genesisData, err := os.ReadFile(genesisFile)
+			if err != nil {
+				return fmt.Errorf("failed to read genesis file: %w", err)
+			}
+
+			// Parse the genesis file
+			var genesis map[string]json.RawMessage
+			err = json.Unmarshal(genesisData, &genesis)
+			if err != nil {
+				return fmt.Errorf("failed to unmarshal genesis file: %w", err)
+			}
+
+			{ // Update block max gas
+				consensusGenesis := &genutiltypes.ConsensusGenesis{}
+				err = consensusGenesis.UnmarshalJSON(genesis["consensus"])
+				if err != nil {
+					return fmt.Errorf("failed to unmarshal consensus genesis: %w", err)
+				}
+
+				if consensusGenesis.Params == nil {
+					consensusGenesis.Params = &cmttypes.ConsensusParams{}
+				}
+				consensusGenesis.Params.Block.MaxGas = 36_000_000
+
+				// Marshal the updated consensus genesis back to genesis
+				updatedConsensusGenesis, err := consensusGenesis.MarshalJSON()
+				if err != nil {
+					return fmt.Errorf("failed to marshal updated consensus genesis: %w", err)
+				}
+				genesis["consensus"] = updatedConsensusGenesis
+			}
+
+			{ // Update the app state
+				var appState map[string]json.RawMessage
+				err = json.Unmarshal(genesis["app_state"], &appState)
+				if err != nil {
+					return fmt.Errorf("failed to unmarshal app state: %w", err)
+				}
+
+				// Update genesis state for each module
+				appState["bank"] = improveGenesisOfBank(appState["bank"], clientCtx.Codec)
+				appState["staking"] = improveGenesisOfStaking(appState["staking"], clientCtx.Codec)
+				appState["mint"] = improveGenesisOfMint(appState["mint"], clientCtx.Codec)
+				appState["evm"] = improveGenesisOfEvm(appState["evm"], clientCtx.Codec)
+				appState["crisis"] = improveGenesisOfCrisis(appState["crisis"], clientCtx.Codec)
+				appState["gov"] = improveGenesisOfGov(appState["gov"], clientCtx.Codec)
+
+				// Marshal the updated app state back to genesis
+				updatedAppState, err := json.Marshal(appState)
+				if err != nil {
+					return fmt.Errorf("failed to marshal updated app state: %w", err)
+				}
+				genesis["app_state"] = updatedAppState
+			}
+
+			// Marshal the updated genesis back to JSON
+			updatedGenesisData, err := json.MarshalIndent(genesis, "", "  ")
+			if err != nil {
+				return fmt.Errorf("failed to marshal updated genesis: %w", err)
+			}
+
+			// Write the updated genesis back to the file
+			err = os.WriteFile(genesisFile, updatedGenesisData, 0o644)
+			if err != nil {
+				return fmt.Errorf("failed to write updated genesis file: %w", err)
+			}
+
+			return nil
+		},
+	}
+
+	return cmd
+}
+
+// improveGenesisOfBank adds denom metadata.
+func improveGenesisOfBank(rawGenesisState json.RawMessage, codec codec.Codec) json.RawMessage {
+	var bankGenesisState banktypes.GenesisState
+	codec.MustUnmarshalJSON(rawGenesisState, &bankGenesisState)
+
+	bankGenesisState.DenomMetadata = append(bankGenesisState.DenomMetadata, banktypes.Metadata{
+		Description: "Native token of the chain",
+		DenomUnits: []*banktypes.DenomUnit{
+			{
+				Denom:    constants.BaseDenom,
+				Exponent: 0,
+			},
+			{
+				Denom:    constants.SymbolDenom,
+				Exponent: constants.BaseDenomExponent,
+			},
+		},
+		Base:    constants.BaseDenom,
+		Display: constants.SymbolDenom,
+		Name:    constants.DisplayDenom,
+		Symbol:  constants.SymbolDenom,
+	})
+
+	return codec.MustMarshalJSON(&bankGenesisState)
+}
+
+// improveGenesisOfStaking updates bond denom.
+func improveGenesisOfStaking(rawGenesisState json.RawMessage, codec codec.Codec) json.RawMessage {
+	var stakingGenesisState stakingtypes.GenesisState
+	codec.MustUnmarshalJSON(rawGenesisState, &stakingGenesisState)
+
+	stakingGenesisState.Params.BondDenom = constants.BaseDenom
+
+	return codec.MustMarshalJSON(&stakingGenesisState)
+}
+
+// improveGenesisOfMint updates mint denom.
+func improveGenesisOfMint(rawGenesisState json.RawMessage, codec codec.Codec) json.RawMessage {
+	var mintGenesisState minttypes.GenesisState
+	codec.MustUnmarshalJSON(rawGenesisState, &mintGenesisState)
+
+	mintGenesisState.Params.MintDenom = constants.BaseDenom
+
+	return codec.MustMarshalJSON(&mintGenesisState)
+}
+
+// improveGenesisOfEvm updates evm denom.
+func improveGenesisOfEvm(rawGenesisState json.RawMessage, codec codec.Codec) json.RawMessage {
+	var evmGenesisState evmtypes.GenesisState
+	codec.MustUnmarshalJSON(rawGenesisState, &evmGenesisState)
+
+	evmGenesisState.Params.EvmDenom = constants.BaseDenom
+
+	return codec.MustMarshalJSON(&evmGenesisState)
+}
+
+// improveGenesisOfCrisis updates crisis denom and fee.
+func improveGenesisOfCrisis(rawGenesisState json.RawMessage, codec codec.Codec) json.RawMessage {
+	var crisisGenesisState crisistypes.GenesisState
+	codec.MustUnmarshalJSON(rawGenesisState, &crisisGenesisState)
+
+	crisisGenesisState.ConstantFee.Denom = constants.BaseDenom
+	crisisGenesisState.ConstantFee.Amount = sdkmath.NewIntFromBigInt(new(big.Int).Exp(
+		big.NewInt(10), big.NewInt(constants.BaseDenomExponent), nil,
+	)).MulRaw(10)
+
+	return codec.MustMarshalJSON(&crisisGenesisState)
+}
+
+// improveGenesisOfGov updates gov params like denom and deposit amount.
+func improveGenesisOfGov(rawGenesisState json.RawMessage, codec codec.Codec) json.RawMessage {
+	var govGenesisState govtypesv1.GenesisState
+	codec.MustUnmarshalJSON(rawGenesisState, &govGenesisState)
+
+	amountOfNative := func(amount int64) sdkmath.Int {
+		return sdkmath.NewIntFromBigInt(new(big.Int).Exp(
+			big.NewInt(10), big.NewInt(constants.BaseDenomExponent), nil,
+		)).MulRaw(amount)
+	}
+
+	govGenesisState.Params.MinDeposit = sdk.NewCoins(sdk.NewCoin(constants.BaseDenom, amountOfNative(1_000)))
+	govGenesisState.Params.ExpeditedMinDeposit = sdk.NewCoins(sdk.NewCoin(constants.BaseDenom, amountOfNative(2_000)))
+
+	return codec.MustMarshalJSON(&govGenesisState)
+}