+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (evm) [Evermint#175](https://github.com/EscanBE/evermint/pull/175) Support precompiled contract using new module `x/cpc`
 - (evm) [Evermint#178](https://github.com/EscanBE/evermint/pull/178) Staking precompiled contract ESIP-179
 - (evm) [Evermint#182](https://github.com/EscanBE/evermint/pull/182) Bech32 precompiled contract ESIP-181
+- (genesis) [#4](https://github.com/EscanBE/everlast/pull/4) Add command to update necessary genesis content
 
 ### Improvement